pyproject.toml
[project]
name = "cot-tampering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff>=0.14.3",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0.0",
    "fire>=0.6.0",
    "jupyter>=1.1.1",
    "tqdm>=4.66.0",
    "numpy>=2.3.4",
    "matplotlib>=3.10.7",
    "transformers>=4.57.1",
    "torch>=2.9.0",
    "omegaconf>=2.3.0",
]
